Definicion experience como principal componente
