+++ cliente/package.json
@@ -19,6 +19,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.11.2",
     "react-scripts": "^5.0.1",
+    "react-three-fiber": "^6.0.13",
     "three": "^0.150.1"
   },
   "browserslist": {
