+++ cliente/src/index2.js
@@ -1,16 +1,8 @@
 import React, { useState } from 'react';
 import { Canvas } from '@react-three/fiber';
-import Scene1 from './Scene1';
-import Scene2 from './Scene2';
 import Experience from './Experience';
 
 function Index2() {
-  const [currentScene, setCurrentScene] = useState('sceneWelcome');
-
-  const changeScene = (scene) => {
-    setCurrentScene(scene);
-  };
-
   return (
     <Canvas
       camera={{
@@ -20,9 +12,7 @@ function Index2() {
         position: [-4, 3, 6],
       }}
     >
-      {currentScene === 'sceneWelcome' && <Experience changeScene={changeScene} />}
-      {currentScene === 'scene1' && <Scene1 changeScene={changeScene} />}
-      {currentScene === 'scene2' && <Scene2 changeScene={changeScene} />}
+      <Experience />
     </Canvas>
   );
 }
